extract singleton metaclass to Utils
