+++ Source/FaceRecognition/Utils/Singleton.py
@@ -1,0 +1,7 @@
+class SingletonMeta(type):
+    _instance = None
+
+    def __call__(self):
+        if self._instance is None:
+            self._instance = super().__call__()
+        return self._instance
